refactor usecase implementation. add get user. delete unused code.
