--- service_notification/src/main/kotlin/org/thechance/service_notification/domain/usecases/CreateUserUseCase.kt
@@ -1,7 +1,0 @@
-package org.thechance.service_notification.domain.usecases
-
-import org.thechance.service_notification.domain.entity.User
-
-interface CreateUserUseCase {
-    suspend operator fun invoke(user:User): Boolean
-}
