--- service_notification/src/main/kotlin/org/thechance/service_notification/domain/usecases/CreateUserUseCaseImp.kt
@@ -1,12 +1,0 @@
-package org.thechance.service_notification.domain.usecases
-
-import org.koin.core.annotation.Single
-import org.thechance.service_notification.domain.entity.User
-import org.thechance.service_notification.domain.gateway.UserGateway
-
-@Single
-class CreateUserUseCaseImp(private val userGateway: UserGateway): CreateUserUseCase {
-    override suspend fun invoke(user: User): Boolean {
-        return userGateway.createUser(user)
-    }
-}
