--- service_notification/src/main/kotlin/org/thechance/service_notification/domain/usecases/GertUsersUseCase.kt
@@ -1,7 +1,0 @@
-package org.thechance.service_notification.domain.usecases
-
-import org.thechance.service_notification.domain.entity.User
-
-interface GertUsersUseCase {
-    suspend operator fun invoke(): List<User>
-}
