+++ service_notification/src/main/kotlin/org/thechance/service_notification/domain/usecases/IUserUseCase.kt
@@ -1,8 +1,8 @@
-package org.thechance.service_notification.domain.gateway
+package org.thechance.service_notification.domain.usecases
 
 import org.thechance.service_notification.domain.model.User
 
-interface UserGateway {
+interface IUserUseCase {
     suspend fun createUser(user: User): Boolean
     suspend fun getUsers(): List<User>
     suspend fun getUser(id:String): User
