--- service_notification/src/main/kotlin/org/thechance/service_notification/domain/usecases/TestUseCase.kt
@@ -1,8 +1,0 @@
-package org.thechance.service_notification.domain.usecases
-
-class TestUseCaseImp : TestUseCase {
-}
-
-interface TestUseCase {
-
-}
