--- service_notification/src/main/kotlin/org/thechance/service_notification/endpoints/dto/Test.kt
@@ -1,2 +1,0 @@
-package org.thechance.service_notification.endpoints.dto
-
